requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
asyncpg==0.28.0
attrs==23.1.0
build==0.10.0
CacheControl==0.12.14
certifi==2023.5.7
charset-normalizer==3.2.0
cleo==2.0.1
click==8.1.6
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.7
dnspython==2.4.0
dulwich==0.21.5
email-validator==2.0.0.post2
fastapi==0.99.1
fastapi-cache2==0.2.1
filelock==3.12.2
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
html5lib==1.1
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
importlib-metadata==6.8.0
installer==0.7.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
keyring==23.13.1
lockfile==0.12.2
Mako==1.2.4
MarkupSafe==2.1.3
more-itertools==10.0.0
msgpack==1.0.5
multidict==6.0.4
orjson==3.9.2
packaging==23.1
pendulum==2.1.2
pexpect==4.8.0
pika==1.3.2
pkginfo==1.9.6
platformdirs==3.9.1
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
ptyprocess==0.7.0
pydantic==1.10.11
pydantic_core==2.3.0
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytzdata==2020.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
rapidfuzz==2.15.1
redis==4.6.0
referencing==0.30.0
requests==2.31.0
requests-toolbelt==1.0.0
rpds-py==0.9.2
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
tomlkit==0.11.8
trove-classifiers==2023.7.6
typing_extensions==4.7.1
ujson==5.8.0
urllib3==1.26.16
uvicorn==0.23.1
virtualenv==20.24.2
watchfiles==0.19.0
webencodings==0.5.1
websockets==11.0.3
yarl==1.9.2
zipp==3.16.2
